+++ api-gateway-service/src/main/java/com/example/apigatewayservice/common/filter/GlobalFilter.java
@@ -1,12 +1,9 @@
-package com.example.apigatewayservice.global.filter;
+package com.example.apigatewayservice.common.filter;
 
-import com.example.apigatewayservice.global.filter.LoggingFilter.Config;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.cloud.gateway.filter.GatewayFilter;
-import org.springframework.cloud.gateway.filter.OrderedGatewayFilter;
 import org.springframework.cloud.gateway.filter.factory.AbstractGatewayFilterFactory;
-import org.springframework.core.Ordered;
 import org.springframework.http.server.reactive.ServerHttpRequest;
 import org.springframework.http.server.reactive.ServerHttpResponse;
 import org.springframework.stereotype.Component;
@@ -14,7 +11,7 @@ import reactor.core.publisher.Mono;
 
 @Slf4j
 @Component
-public class LoggingFilter extends AbstractGatewayFilterFactory<Config> {
+public class GlobalFilter extends AbstractGatewayFilterFactory<GlobalFilter.Config> {
 
     @Data
     public static class Config {
@@ -23,36 +20,29 @@ public class LoggingFilter extends AbstractGatewayFilterFactory<Config> {
         public boolean postLogger;
     }
 
-    public LoggingFilter() {
+    public GlobalFilter() {
         super(Config.class);
     }
 
     @Override
     public GatewayFilter apply(Config config) {
 
-        /**
-         * [ OrderedGatewayFilter ]
-         * - 필터의 우선순위를 지정하기 위한 구현체
-         */
-        GatewayFilter filter =  new OrderedGatewayFilter((exchange, chain) -> {
+        return ((exchange, chain) -> {
             ServerHttpRequest request = exchange.getRequest();
             ServerHttpResponse response = exchange.getResponse();
 
-            log.info("Logging Filter Message: {}", config.getBaseMessage());
+            log.info("Global Filter Message: {}", config.getBaseMessage());
 
             if (config.isPreLogger()) {
-                log.info("Logging PRE Filter: request id -> {}", request.getId());
+                log.info("Global Filter Start: request id -> {}", request.getId());
             }
 
             return chain.filter(exchange).then(Mono.fromRunnable(() -> {
-                log.info("Logging POST Filter: response code -> {}", response.getStatusCode());
+                log.info("Global Filter End: response code -> {}", response.getStatusCode());
             }));
 
-        }, Ordered.LOWEST_PRECEDENCE);
-
-        return filter;
+        });
 
     }
-
 
 }
